CQRS - WIP

+++ st_server/shared/helper/file.py
@@ -1,20 +1,18 @@
 """File helper functions."""
 
-import configparser
 import os
 
 
-def save(file: bytes, filename: str) -> str:
+def save(file: bytes, filepath: os.PathLike, overwrite: bool = False) -> None:
     """Saves a file to the storage path."""
-    config = configparser.ConfigParser()
-    config.read("st_server/config.ini")
-    base_path = config.get("storage", "path")
-    filepath = os.path.join(base_path, filename)
+    os.makedirs(os.path.dirname(filepath), exist_ok=True)
+    if os.path.isfile(filepath) and not overwrite:
+        raise FileExistsError
     with open(filepath, "wb") as f:
         f.write(file)
-    return filepath
 
 
 def delete(filepath: str) -> None:
     """Deletes a file from the storage path."""
-    os.remove(filepath)
+    if os.path.isfile(filepath):
+        os.remove(filepath)
